Development

+++ client/src/components/Add.js
@@ -1,14 +1,24 @@
-import React from 'react';
-import '../styles/Podcast.css';
+import React, { useEffect, useState } from 'react';
+// import '../styles/Podcast.css';
 import { IoIosAddCircle } from 'react-icons/io';
 
 const Add = ({ list, setList, obj }) => {
+  const alreadyExists = list.find(element => element.id === obj.id);
+  // const [alreadyAdded, setAdded] = useState(alreadyExists ? true : false);
+  
+  // useEffect(() => {
+  //   if(alreadyExists) {
+  //     setAdded(true)
+  //   }
+  //   console.log(alreadyExists)
+  //   console.log(list)
+  // }, []);
+
+  const added = "podcast__button--add " + (alreadyExists ? "added" : ""); 
+
   const addToList = () => {
-    // const alreadyExists = list.indexOf(obj);
-    // if(alreadyExists !== -1) {
-    //   window.alert('it is already in the list')
-    //   return;
-    // }
+    console.log(alreadyExists, obj)
+    if (alreadyExists !== undefined) return;
     list.push(obj);
     setList(list);
     localStorage.setItem('list', JSON.stringify(list));
@@ -16,7 +26,7 @@ const Add = ({ list, setList, obj }) => {
 
   return (
     <>
-      <IoIosAddCircle onClick={addToList} className="podcast__button--icon podcast__button--add"/>
+      <IoIosAddCircle onClick={addToList} className={added}/>
     </>
   )
 }
